Initial MSVC support

+++ src/bot_api/types.h
@@ -103,7 +103,7 @@ public:
   // Returns the number of moves until checkmate in the format described above. If `type()` is not
   // equal to `Centipawns`, then the behaviour is undefined
   inline constexpr int16_t checkMate() const {
-    return (value_ < 0) ? (VALUE_MIN - value_) : (VALUE_MAX - value_ + 1);
+    return static_cast<uint16_t>((value_ < 0) ? (VALUE_MIN - value_) : (VALUE_MAX - value_ + 1));
   }
 
   // Comparison operators
